pyproject.toml
[project]
name = "web-search-gym"
version = "0.1.0"
description = "Training gym for web-search agents. SFT + GRPO on Qwen3 with Exa search, evaluated on BrowseComp."
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "rapidfuzz>=3.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
agents = [
    "anthropic>=0.30",
    "openai>=1.0",
    "exa-py>=1.0",
]
bench = [
    "tqdm>=4.65",
]
sft = [
    "torch>=2.1",
    "transformers>=4.51",
    "trl>=1.0",
    "peft>=0.14",
    "datasets>=3.0",
    "accelerate>=1.0",
]
rl = [
    # Populated when the RL phase lands. Expected: trl, transformers,
    # peft, accelerate (same family as SFT), plus wandb for logging.
]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
wsg = "bench.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "bench", "sft", "synth"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
